Temporarily replaced calls to Solo.waitForDialogToClose() with Instrumentation.waitForIdlySync().
There appears to be a race condition in waitForDialogToClose().  It
first calls waitForDialogToOpen() before waiting for the dialog to
close.  If the dialog is already closed by this time (which appears to
be happening in most of my runs), it will add an additional "timeout"
milliseconds to the run time of each test.

I don't know if waitForIdleSync() is actually doing what I intended,
namely waiting for the dialog to close, but it appears to work on tests
run on my machine.